add order func
