--- base/serializers/orders.py
@@ -1,29 +1,0 @@
-from django.contrib.auth.models import User
-from drf_yasg.utils import swagger_serializer_method
-from rest_framework import serializers
-from rest_framework_simplejwt.tokens import RefreshToken
-
-# from backend.base.models import UserProfile
-from base.models import Order, Review
-from base.serializers.users import *
-
-
-class ReviewSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = Review
-        fields = '__all__'
-
-
-class OrderSerializer(serializers.ModelSerializer):
-    user = serializers.SerializerMethodField(read_only=True)
-
-    class Meta:
-        model = Order
-        fields = '__all__'
-
-    def get_user(self, obj):
-        user = obj.user
-        serializer = TraineeSerializer(user, many=False)
-        return serializer.data
-
-
